Compute Confidence Intervals for an lmFScreen Model

Prints and returns confidence intervals for both selective and naive estimates. The confidence level is taken from the alpha value stored in the lmFScreen object.

Usage

# S3 method for class 'lmFScreen'
confint(object, ...)

Arguments

object

An object of class lmFScreen containing model results.

...

Additional arguments (currently unused).

Value

Invisibly returns a list containing:

Selective: A data frame of confidence intervals for selective estimates

Naive: A data frame of confidence intervals for naive estimates

Examples

data(mtcars)
result <- lmFScreen(mpg ~ wt + hp, data = mtcars)
confint(result)
#> 
#> lmFScreen Model Confidence Intervals
#> ------------------------------------------------------
#> Confidence Level: 95.00%
#> Number of predictors: 2
#> ------------------------------------------------------
#> 
#> Selective Confidence Intervals:
#> Predictor       Lower.CI     Upper.CI
#> ---------------------------------------------
#>  wt                -5.1687       -2.5843
#>  hp                -0.0502       -0.0133
#> 
#> Naive Confidence Intervals:
#> Predictor       Lower.CI     Upper.CI
#> ---------------------------------------------
#>  wt                -5.1180       -2.6377
#>  hp                -0.0495       -0.0141
#>